Add files via upload
This is a ppt describing NumPy and its real time applications. NumPy library provide many modules to operate on large multi dimensional arrays and matrices.
In python lists serve in place of arrays, but it takes time to process. 
NumPy arrays are stored at one continuous place in memory unlike lists, so processes can access and manipulate them very efficiently.
This behavior is called locality of reference in computer science.
This is the main reason why NumPy is faster than lists. Also it is optimized to work with latest CPU architectures.